fixed checkstyle & pmd
